Create README.md
Having a health insurance can be expensive but it's a necessary when you can afford it. Here we are analyzing the insurance data set from four regions of US. The main objective is to create a model that can predict medical insurance charges with greater accuracy and find out factors that heavily influence the rising cost.